feat(purchase): add purchase functionality
- add tests for purchase functionality
- add model column for price

[Delivers #163483849]

+++ src/db/models/transaction.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
     buyerId: DataTypes.INTEGER,
     sellerId: DataTypes.INTEGER,
     artId: DataTypes.INTEGER,
-    amount: DataTypes.INTEGER
+    amount: DataTypes.DECIMAL(10, 2),
   }, {});
   Transaction.associate = function (models) {
     // associations can be defined here
